pyproject.toml
[project]
name = "slack-airtable-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "jupyter>=1.1.1",
    "pyairtable>=3.1.1",
    "python-dotenv>=1.1.0",
    "slack-sdk>=3.35.0",
]
